refactor: extract logic from HandleProjectRetrieved.php to ProjectProcessingService.php
